requirements.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.2
automium==0.2.6
automium-web==0.1.1
autopep8==1.4.4
bcrypt==3.2.0
blinker==1.4
CacheControl==0.12.6
cachetools==4.2.0
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
Click==7.0
click-default-group==1.2.2
colorama==0.4.3
Contacts==1.0
contextlib2==0.6.0.post1
cryptography==2.8
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-allauth==0.43.0
django-cors-headers==3.6.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-extensions==3.0.2
django-heroku==0.3.1
django-livesync==0.5
django-tinymce==3.2.0
django-tinymce4-lite==1.8.0
djangorestframework==3.12.2
docopt==0.6.2
docutils==0.16
entrypoints==0.3
filelock==3.0.12
firebase-admin==4.5.1
Flask==1.1.2
future==0.18.2
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
graphviz==0.16
grpcio==1.35.0
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
importlib==1.0.4
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
jsmin==2.2.2
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
legacy==0.1.6
livereload==2.6.3
lockfile==0.12.2
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
msgpack==1.0.2
nodeenv==1.5.0
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pathtools==0.1.2
pbr==5.5.1
pep8==1.7.1
pexpect==4.6.0
Pillow==7.0.0
pipenv==2020.11.15
proto-plus==1.13.0
protobuf==3.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pydotplus==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.7.3
python3-openid==3.2.0
pythonanywhere==0.9.5
pytube==9.6.0
pytz==2019.3
pywin32-ctypes==0.2.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7
schema==0.7.2
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
tabulate==0.8.7
task==0.2.0
testresources==2.0.1
toml==0.10.1
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.8
views==0.3
virtualenv==20.1.0
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wadllib==1.3.3
waitress==1.4.4
watchdog==0.10.3
websocket-client==0.57.0
Werkzeug==1.0.1
whitenoise==4.1.2
wrapt==1.12.1
